% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_minimum_distance.R
\name{average_minimum_distance}
\alias{average_minimum_distance}
\title{average_minimum_distance}
\usage{
average_minimum_distance(spe_object)
}
\arguments{
\item{spe_object}{SpatialExperiment object in the form of the output of
\code{\link{format_image_to_spe}}.}
}
\value{
A single number is returned
}
\description{
Calculates the average minimum distance of all cells to their
  nearest cells in the input image.
}
\examples{
average_minimum_distance(SPIAT::simulated_image)
}
