% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_nearest_neighbor_index.R
\name{average_nearest_neighbor_index}
\alias{average_nearest_neighbor_index}
\title{ANN index for point pattern (clustering or dispersion)}
\usage{
average_nearest_neighbor_index(
  sce_object,
  reference_cell,
  feature_colname,
  p.val = 5e-06
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of
format_image_to_sce.}

\item{reference_cell}{String. Cells of this type are used as reference cells.}

\item{feature_colname}{String. Specify the selected column for
`reference_cell`.}

\item{p.val}{Numeric. The p value threshold to determine the significance of
a pattern.}
}
\description{
Calculate the the ANN index of a specified type of cells. The
  index indicates the clustering effect of a point pattern. It can be
  clustering, random or dispersion.
}
