% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_dependence.R
\name{bivariate_dependence}
\alias{bivariate_dependence}
\title{bivariate_dependence}
\usage{
bivariate_dependence(
  sce_object,
  method = "Kcross",
  phenotypes,
  merge = NULL,
  merge_name = NULL
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{method}{String that is the method for dependence calculation}

\item{phenotypes}{Vector of phenotypes of interest}

\item{merge}{Vector of phenotypes to be merged}

\item{merge_name}{String that is the name of the merged phenotype}
}
\description{
find the dependence between two types of cells
}
