% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_all_distances_between_cell_types.R
\name{calculate_all_distances_between_cell_types}
\alias{calculate_all_distances_between_cell_types}
\title{calculate_all_distances_between_cell_types}
\usage{
calculate_all_distances_between_cell_types(
  sce_object,
  cell_types_of_interest = NULL,
  column
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{cell_types_of_interest}{Vector containing cell types to be considered,
if NULL, all cell type combinations will be calculated}

\item{column}{Column name with the cell types of interest to be considered}
}
\value{
A data.frame is returned.
}
\description{
Returns the distances between cells of different types. 
If none of the cell types are found, it will print an error message and return a vector of NAs
}
