% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_all_distances_between_celltypes.R
\name{calculate_all_distances_between_celltypes}
\alias{calculate_all_distances_between_celltypes}
\title{calculate_all_distances_between_celltypes}
\usage{
calculate_all_distances_between_celltypes(
  sce_object,
  remove_other = TRUE,
  cell_types_of_interest = NULL
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{remove_other}{If TRUE, the plotting will exclude undefined cells}

\item{cell_types_of_interest}{Vector containing cell types to be considered,
if NULL, all cell type combinations will be calculated}
}
\value{
A data.frame is returned.
}
\description{
Returns the distances between cells of different cell types
}
\examples{
distances <- calculate_all_distances_between_celltypes(SPIAT::formatted_image, remove_other = TRUE, cell_types_of_interest = c("CD3,CD4", "CD3,CD8"))
}
