% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_all_distances_between_phenotypes.R
\name{calculate_all_distances_between_phenotypes}
\alias{calculate_all_distances_between_phenotypes}
\title{calculate_all_distances_between_phenotypes}
\usage{
calculate_all_distances_between_phenotypes(
  sce_object,
  remove_other = TRUE,
  cell_phenotypes_of_interest = NULL
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{remove_other}{If TRUE, the plotting will exclude cells with only DAPI marker}

\item{cell_phenotypes_of_interest}{Vector containing phenotypes to be considered,
if NULL, all phenotype combinations will be calculated}
}
\description{
Returns the distances between cells of different phenotypes
}
\examples{
distances <- calculate_all_distances_between_phenotypes(SPIAT::formatted_image, remove_other = TRUE, cell_phenotypes_of_interest = c("CD3,CD4", "CD3,CD8"))
}
