% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cell_proportions.R
\name{calculate_cell_proportions}
\alias{calculate_cell_proportions}
\title{calculate_cell_proportions}
\usage{
calculate_cell_proportions(sce_object)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}
}
\description{
Calculate the number and proportion of each cell phenotype in image
}
