% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cross_k_function.R
\name{calculate_cross_k_function}
\alias{calculate_cross_k_function}
\title{Perform cross-K function}
\usage{
calculate_cross_k_function(
  point_pattern,
  phenotype_group1,
  phenotype_group2,
  names = NULL,
  plot = 1
)
}
\arguments{
\item{point_pattern}{ppp object representing image}

\item{phenotype_group1}{Cell phenotypes to be compared from, can be a single phenotype or multiple}

\item{phenotype_group2}{Cell phenotypes to be compared to, can be a single phenotype or multiple}

\item{names}{String vector containing the names of the phenotype groups}

\item{plot}{Dictate whether to plot Cross-K function}
}
\description{
Calculate Cross-K function for 2 groups of cell phenotypes of interest
}
\details{
calculate_cross_k_function, Calculates Cross-K function between 2 specified cell phenotype groups.
The phenotype groups can contain multiple or single phenotypes. Options are given to name the two groups and to plot the function.

Inputs - 
point_pattern: marked ppp object representing intial image
phenotype_group1: String containing the names of the phenotypes to be considered, can be singular or multiple.
phenotype_group2: See phenotype_group_1
plot: Indicates whether to plot Cross-K function by checking for numeric value. By default set to plot graph (plot = 1), if no plot required set to NULL
names: String vector containing desired names for the two groups of phenotypes, largely for plotting purposes. If not specified groups are named Group1 and Group2.
}
