% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cross_k_multiple_images.R
\name{calculate_cross_k_multiple_images}
\alias{calculate_cross_k_multiple_images}
\title{Calculate K-cross function for multiple images and plot}
\usage{
calculate_cross_k_multiple_images(
  image_point_patterns,
  phenotype_group1,
  phenotype_group2,
  names = NULL,
  plot = 2
)
}
\arguments{
\item{image_point_patterns}{Point patterns to be investigated}

\item{phenotype_group1}{Names of phenotypes of interest (Group1)}

\item{phenotype_group2}{Names of phenotypes of interest (Group2)}

\item{names}{String vector containing the names of the phenotype groups}

\item{plot}{Dictate whether plotting of function is required}
}
\description{
Inputs -
image_point_patterns: List containing point pattern objects of the images of interest
phenotype_group1: Names of phenotypes of interest for group 1
phenotype_group2: Names of phenotypes of interest for group 2
names: String vector containing desired names for the two groups of phenotypes, largely for plotting purposes. If not specified groups are named Group1 and Group2.
plot: Dictates whether to plot cross K-functions calculated, set to NULL if plotting is not required
}
\details{
Outputs - 
k_cross_functions: list containing the data required to plot multiple cross-K functions
}
