% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_entropy.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{calculate_entropy}
\usage{
calculate_entropy(
  sce_object,
  cell_types_of_interest,
  feature_colname = "Phenotype",
  radius = NULL
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{cell_types_of_interest}{Cell.Types of interest, the first celltype is considered as reference celltype}

\item{feature_colname}{String specifying the column the markers are from}

\item{radius}{(OPTIONAL) The maximum radius around a reference cell for another cell to be considered an interaction.}
}
\value{
A dataframe or a number depending on the argument radius
}
\description{
Produces a dataframe of entropies for all reference cells or an entropy number for the whole image if a radius is not supplied.
}
