% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_k_cross_function.R
\name{calculate_k_cross_function}
\alias{calculate_k_cross_function}
\title{Perform cross-K function}
\usage{
calculate_k_cross_function(point_pattern, cell_phenotypes_of_interest = NULL)
}
\arguments{
\item{point_pattern}{ppp object representing image}

\item{cell_phenotypes_of_interest}{cell phenotypes to be compared}
}
\description{
calculate cross-K function for 2 chosen phenotypes of interest
}
