% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_min_distances_to_borders.R
\name{calculate_min_distance_to_borders}
\alias{calculate_min_distance_to_borders}
\title{calculate_min_distance_to_borders}
\usage{
calculate_min_distance_to_borders(sce_object, split = NULL)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{split}{integer specifying the number of subimages to split; if NULL, do not split}
}
\value{
An sce_object is returned
}
\description{
Returns the sce_object with the minimum distance from immune cells to the identified bordering cells
}
