% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_minimum_distances_between_cell_types.R
\name{calculate_minimum_distances_between_cell_types}
\alias{calculate_minimum_distances_between_cell_types}
\title{calculate_minimum_distances_between_cell_types}
\usage{
calculate_minimum_distances_between_cell_types(
  sce_object,
  feature_colname = "Phenotype",
  all_combinations = FALSE,
  cell_types_of_interest
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of
format_image_to_sce.}

\item{feature_colname}{String. Column of cells to choose the cell type from
(e.g. Cell.Type, Cell.Type2, etc).}

\item{all_combinations}{Boolean. If TRUE, the distances between all possible
combinations of cell types will be calculated.}

\item{cell_types_of_interest}{String Vector of cell types to consider if
all_combinations is FALSE.}
}
\value{
A data.frame is returned
}
\description{
Returns the mean, median and stardard deviation of the minimum
  distances between phenotypes.
}
\examples{
min_distances <- 
calculate_minimum_distances_between_cell_types(SPIAT::defined_image,
feature_colname = "Cell.Type", cell_types_of_interest = c("Tumour","Immune1"))
}
