% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spatial_autocorrelation.R
\name{calculate_spatial_autocorrelaiton}
\alias{calculate_spatial_autocorrelaiton}
\title{calculate_spatial_autocorrelation}
\usage{
calculate_spatial_autocorrelaiton(raster_obj, metric = "globalmoran")
}
\arguments{
\item{raster_obj}{Raster object in the form of the output of raster function.}

\item{metric}{String. The method for calculating spatial autocorrelation.
Choose from "globalmoran" and "GearyC".}
}
\description{
Calculate the spatial autocorrelation metric of the grid squares
  that have a certain pattern.
}
