% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summary_distances_to_borders.R
\name{calculate_summary_distances_to_borders}
\alias{calculate_summary_distances_to_borders}
\title{calculate_summary_distances_to_borders}
\usage{
calculate_summary_distances_to_borders(
  sce_object,
  types_of_interest = NULL,
  column = NULL
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{types_of_interest}{Vector of cell types to consider}

\item{column}{String specifying which column the interested cell types are from}
}
\value{
A data.frame is returned
}
\description{
Returns the mean, median and stardard deviation of the distances between a specified cell type to the borders
}
