% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_communities_morisita_index.R
\name{clusters_communities_morisita_index}
\alias{clusters_communities_morisita_index}
\title{clusters_communities_morisita_index}
\usage{
clusters_communities_morisita_index(
  formatted_data_with_clusters_or_communities,
  column_to_consider,
  clusters_or_communities_of_interest = NULL,
  CI = 0.95
)
}
\arguments{
\item{formatted_data_with_clusters_or_communities}{A dataframe output from identify_cell_clusters or identify_cell_communities}

\item{column_to_consider}{Column name to consider as community/clusters}

\item{clusters_or_communities_of_interest}{A vector containing the names of clusters/communities to calculate morisita index}

\item{CI}{Confidence interval for calculating morisita index}
}
\description{
Calculate the morisita index between different clusters or communities
}
