% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colhex2col.R
\name{colhex2col}
\alias{colhex2col}
\title{colhex2col}
\usage{
colhex2col(colhex)
}
\arguments{
\item{colhex}{- A vector of hexadecimal colours.}
}
\value{
the human readable colour converted last from \code{colhex}.
}
\description{
Converts a set of hexadecimal colours into normal human readable colours
}
