% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_of_clusters_and_communities.R
\name{composition_of_clusters_and_communities}
\alias{composition_of_clusters_and_communities}
\title{composition_of_clusters_and_communities}
\usage{
composition_of_clusters_and_communities(
  formatted_data_with_clusters,
  column_to_consider
)
}
\arguments{
\item{formatted_data_with_clusters}{- a dataframe output from generate_clusters}

\item{column_to_consider}{Column name to consider as community/clusters}
}
\description{
Returns a dataframe which contains the percentages of cells with a
specific marker within each cluster and the number of cells in the cluster.
}
