% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_celltypes.R
\name{define_celltypes}
\alias{define_celltypes}
\title{define_celltypes}
\usage{
define_celltypes(sce_object, phenotypes = NULL, names = NULL)
}
\arguments{
\item{sce_object}{SingleCellExperiment object generated by format_image_to_sce}

\item{names}{Vector of new names assigned to the selected phenotypes; if NULL, 
the function will use predefined phenotypes and names}

\item{phentoypes}{Vector of phenotypes of keep or exclude; if NULL, the function will use 
predefined phenotypes and names}
}
\description{
Define celltypes of selected phenotypes. If the specified phenotype vector is 
shorter than the phenotypes in the sce_object, then the unspecified phenotypes will be defined
as "Undefined"
}
