% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_phenotypes.R
\name{define_phenotypes}
\alias{define_phenotypes}
\title{define_phenotypes}
\usage{
define_phenotypes(sce_object, phenotypes = NULL, names = NULL)
}
\arguments{
\item{sce_object}{SingleCellExperiment object generated by format_image_to_sce}

\item{phenotypes}{Vector of phenotypes of keep or exclude}

\item{names}{Vector of new names assigned to the selected phenotypes}
}
\description{
Define phenotypes of selected cells
}
