% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_structure.R
\name{define_structure}
\alias{define_structure}
\title{define_structure}
\usage{
define_structure(sce_object, names_of_immune_cells, n_invasive = 5)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{names_of_immune_cells}{vector indicating the names of potential immune cells}

\item{n_invasive}{Number specifying the number of cells that are included in the invasive front}
}
\description{
Identify the cells that compose the invasive front, the infiltration and the exclusion
}
