% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defined_image.R
\docType{data}
\name{defined_image}
\alias{defined_image}
\title{SCE object of a simulated image with defined cell types based on marker
combinations.}
\format{
An sce object. Assay contains 5 rows (markers) and 4951 columns
  (cells); colData contains 4951 rows (cells) and 4 columns (metadata).
}
\usage{
defined_image
}
\description{
A dataset that contains a formatted sce object  \code{\link{simulated_image}}
with cell ids, cell positions, phenotypes, defined cell types in metadata and
marker intensities in assays.
}
\seealso{
\code{\link{simulated_image}} \code{\link{image_no_markers}}
}
\keyword{datasets}
