% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionality_reduction_plot.R
\name{dimensionality_reduction_plot}
\alias{dimensionality_reduction_plot}
\title{Dimensionality reduction plot}
\usage{
dimensionality_reduction_plot(
  sce_object,
  plot_type = "UMAP",
  scale = TRUE,
  feature_colname
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce.}

\item{plot_type}{String. Choose from "UMAP" and "TSNE".}

\item{scale}{Boolean. Whether scale the marker intensities.}

\item{feature_colname}{String. Specify the column name to group the cells.}
}
\value{
A plot
}
\description{
Generates the dimensionality reduction plots (UMAP or tSNE)
  based on marker intensities. Cells are grouped by the categories under the
  selected column. -- have you tried doing PCA on the matrix and then doing the UMAP/tSNE? Does it help? scRNAseq workflows do this.
}
\examples{
dimensionality_reduction_plot(SPIAT::simulated_image, plot_type = "TSNE", 
feature_colname = "Phenotype")
}
