% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_gradient_aggregated.R
\name{entropy_gradient_aggregated}
\alias{entropy_gradient_aggregated}
\title{The aggregated gradient of entropy and the peak of the gradient}
\usage{
entropy_gradient_aggregated(
  sce_object,
  cell_types_of_interest,
  feature_colname,
  radii
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object.}

\item{cell_types_of_interest}{String Vector. The cell types that the entropy
is computed upon. The reference cell.}

\item{feature_colname}{String. The column name of the interested cell types.}

\item{radii}{Numeric Vector. A vector of radii within a circle of a reference
cell where the entropy is computed on.}
}
\value{
A list of the gradient of entropy and the peak
}
\description{
This function first calculates the entropy of interested cell
  types within a circle of each reference cell at each radii. Then at each
  radii, the entropies of all circles are aggregated into one number. At the
  end, the function returns the aggregated entropy at all radii.
}
