% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_subset.R
\name{find_subset}
\alias{find_subset}
\title{Determine point subset data}
\usage{
find_subset(point_pattern, cell_phenotypes_of_interest)
}
\arguments{
\item{point_pattern}{ppp object representing initial image}

\item{cell_phenotypes_of_interest}{Phenotype/Phenotypes of interest}
}
\description{
find_subset returns unmarked point pattern representing the phenotype/phenotypes 
of interest from the initial image
}
\details{
Inputs - 
point_pattern: Marked point pattern representing data
cell_phenotypes_of_interest: String representation of the names of the phenotypes of interest

Outputs - 
subset_of_interest: unmarked ppp object which is subset of points representing the phenotype specified,
}
