% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cluster_model.R
\name{fit_cluster_model}
\alias{fit_cluster_model}
\title{Fit Cauchy model to point pattern}
\usage{
fit_cluster_model(point_pattern, phenotypes_of_interest = NULL)
}
\arguments{
\item{point_pattern}{ppp object representing data, can be marked point pattern of whole sample or of a single phenotype}

\item{phenotypes_of_interest}{specifies which phenotype to base cluster model}
}
\description{
fit_cluster_model fits a Cauchy cluster model to the data given in point_pattern. 
If phenotypes_of_interest is given only a subset of data will be used to create the model.
}
\details{
Inputs - 
point_pattern: ppp object representing data can be marked or unmarked
phenotypes_of_interest: String with names of phenotypes of interest, if specified the model will be produced based off the subset data. Set to NULL by default

Output - 
model_parameters: List containing model parameters, kappa, mean cluster size and estimated intensity
}
