% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cluster_model_multiple_images.R
\name{fit_cluster_model_multiple_images}
\alias{fit_cluster_model_multiple_images}
\title{Find and plot cluster model parameters of multiple images in a boxplot}
\usage{
fit_cluster_model_multiple_images(
  image_point_patterns = NULL,
  phenotypes_of_interest = NULL,
  plot = 1
)
}
\arguments{
\item{image_point_patterns}{list containing point pattern objects to be compared}

\item{phenotypes_of_interest}{string vector containing name of cell phenotype of interest, can be a single or multiple phenotypes. 
By default it is not specified}

\item{plot}{variable to indicate whether to plot the cluster model data, default is to plot, set to 
NULL if plot is not required}
}
\description{
fit_cluster_model_multiple_images takes in a list containing image point patterns, specific phenotypes
may be specified. The function outputs a list containing vectors containing kappa, estimated point intensity and mean
cluster size. The function by default displays the data as boxplots
}
\details{
Inputs - 
image_point_patterns: list of ppp objects to be analysed
plot: Parameter to indicate whether plotting of data is necessary 
phenotypes_of_interest: Cell phenotypes to be considered when fitting the cluster model
}
