% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_colData_to_sce.R
\name{format_colData_to_sce}
\alias{format_colData_to_sce}
\title{format_colData_to_sce}
\usage{
format_colData_to_sce(df)
}
\arguments{
\item{df}{Dataframe that will be the colData of the sce object.}
}
\value{
An SingleCellExperiment object
}
\description{
Format a dataframe of colData into a singlecellexperiment class
  where the count assay is empty every cell (columns), and cell phenotype, x
  and y coordinates are stored under colData for the purpose of passing
  dataframe into a function requiring sce_object.
}
\examples{
df <- data.frame(Cell.ID = c("Cell_1", "Cell_2"), Cell.X.Positions = c(2,5), Cell.Y.Positions = c(3.3, 8),
Phenotypes = c("CD3", "CD3,CD8"))
sce <- format_colData_to_sce(df)
}
