% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_image_to_sce.R
\name{format_image_to_sce}
\alias{format_image_to_sce}
\title{format_image_to_sce}
\usage{
format_image_to_sce(
  format = "INFORM",
  image,
  markers,
  dye_columns_interest = NULL,
  intensity_columns_interest
)
}
\arguments{
\item{format}{- a string: either "INFORM" or "HALO"}

\item{image}{- a string of the location of either HALO csv file or INFORM textfile}

\item{markers}{- a vector containing the markers used for staining}

\item{dye_columns_interest}{- only for HALO formats, it is a vector of length equal to
markers which contains the column names for marker expression status, either 0
for no expression or 1 for expression. it must be in the same order as the markers}

\item{intensity_columns_interest}{- a vector of length equal to the markers which
contains the column names for the expression level of each of the markers.
It must be in the same order as the markers specified.}
}
\description{
Formats an INFORM or HALO image into a singlecellexperiment class
where the count assay stores the expression level of every marker (rows) for
every cell (columns), and cell phenotype, x and y coordinates are stored
under colData
}
