% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cellprofiler_to_spe.R,
%   R/format_image_to_spe.R
\name{format_image_to_spe}
\alias{format_image_to_spe}
\title{Format an image into a SpatialExperiment object}
\usage{
format_image_to_spe(intensity_matrix, phenotypes = NULL, coord_x, coord_y)

format_image_to_spe(intensity_matrix, phenotypes = NULL, coord_x, coord_y)
}
\arguments{
\item{intensity_matrix}{A matrix of marker intensities or gene expression
where the column names are the Cell IDs, and the rownames the marker.}

\item{phenotypes}{(Optional) String Vector of cell phenotypes in the same
order in which they appear in `intensity_matrix`. If no phenotypes
available, then a vector of NAs can be used as input. Note that the
combination of markers (e.g. CD3,CD4) needs to be used instead of the cell
type name (e.g. helper T cells).}

\item{coord_x}{Numeric Vector with the X coordinates of the cells. The cells
must be in the same order as in the `intensity_matrix`.}

\item{coord_y}{Numeric Vector with the Y coordinates of the cells. The cells
must be in the same order as in the `intensity_matrix`.}

\item{path}{String of the path location cellprofiler csv file.}

\item{markers}{String Vector containing the markers used for staining.}

\item{intensity_columns_interest}{String Vector with the names of the columns
with the level of each marker. Column names must match the order of the
'markers' parameter.}
}
\value{
A SingleCellExperiment object is returned

A SpatialExperiment object is returned
}
\description{
Reads in spatial data in the form of cell coordinates, cell
  phenotypes (if available), and marker intensities and transforms to a
  SingleCellExperiment object. Users can input data in a general format
  (recommended), or data generated from inForm, HALO, Visum, CODEX,
  cellprofiler. The count assay stores the intensity level of every marker
  (rows) for every cell (columns). Cell phenotype, x and y coordinates and
  other cell properties are stored under colData. For the INFORM format, the
  cell properties needed to be included are Cell.Area, Nucleus.Area,
  Nucleus.Compactness, Nucleus.Axis.Ratio, and Cell.Axis.Ratio. For the HALO
  format, the cell properties to be included are Cell.Area, Nucleus.Area and
  Cytoplasm.Area. Note that if the data does not include these parameters, we
  recommend adding it to the output from inForm or HALO with NAs in columns,
  or alternatively using the ‘generic formatting’ option.

Reads in spatial data in the form of cell coordinates, cell
  phenotypes (if available), and marker intensities and transforms to a
  SpatialExperiment object. The assay stores the intensity level of every
  marker (rows) for every cell (columns). Cell phenotype is stored under
  `colData()`. Cell x and y coordinates are stored under `spatialCoords()`
  field.
}
\examples{
#Construct a marker intensity matrix (rows are markers, columns are cells)
intensity_matrix <- matrix(c(14.557, 0.169, 1.655, 0.054, 17.588, 0.229,
1.188, 2.074, 21.262, 4.206,  5.924, 0.021), nrow = 4, ncol = 3)
# define marker names as rownames
rownames(intensity_matrix) <- c("DAPI", "CD3", "CD4", "AMACR")
# define cell IDs as colnames
colnames(intensity_matrix) <- c("Cell_1", "Cell_2", "Cell_3")
# Construct a dummy metadata (phenotypes, x/y coordinates)
# the order of the elements in these vectors correspond to the cell order
# in `intensity matrix`
phenotypes <- c("OTHER",  "AMACR", "CD3,CD4")
coord_x <- c(82, 171, 184)
coord_y <- c(30, 22, 38)

formatted_image <- format_image_to_spe(intensity_matrix=intensity_matrix,
phenotypes = phenotypes, coord_x = coord_x,coord_y = coord_y)
}
