% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted_image.R
\docType{data}
\name{formatted_image}
\alias{formatted_image}
\title{Example Inform image}
\format{
A SingleCellExperiment object where the count assay stores the expression level of every marker (rows) for
}
\usage{
data(formatted_image)
}
\description{
A dataset containing an example Inform image, loaded and formatted as described in data-raw/formatted_image.R
}
\keyword{datasets}
