% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{gg_color_hue}
\usage{
gg_color_hue(number_of_colours)
}
\arguments{
\item{number_of_colours}{- The number of hexadecimal colours to generate.}
}
\value{
A vector of hexadecimal colours
}
\description{
Generates a vector of colours in hexadecimal
}
