% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_bordering_cells_interactive.R
\name{identify_bordering_cells_interactive}
\alias{identify_bordering_cells_interactive}
\title{identify_bordering_cells_interactive}
\usage{
identify_bordering_cells_interactive(
  sce_object,
  reference_cell,
  n_of_polygons = 1,
  buffer_width = 20,
  ahull_alpha = NULL,
  large = FALSE
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{reference_cell}{Cells positive for this marker will be used as reference}

\item{n_of_polygons}{Number specifying the number of tumour regions defined by user}

\item{buffer_width}{Number specifying the allowed margin of error}

\item{ahull_alpha}{Number specifying the ahull parameter. Larger number, more points included in the ahull.}

\item{large}{Boolean specifying if the image requires splitting}
}
\description{
Identify the cells bordering a group of cells of a particular phenotype
}
