% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_cell_communities.R
\name{identify_cell_communities}
\alias{identify_cell_communities}
\title{identify_cell_communities}
\usage{
identify_cell_communities(
  sce_object,
  clustering_method = "dbscan",
  radius = NULL,
  min_community_size = 50,
  phenotypes_of_interest = NULL
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{clustering_method}{String specifying which clustering algorithm to use. Current options:
"dbscan" and "rphenograph"}

\item{radius}{Integer specifying the radius of search. Required for "dbscan"}

\item{min_community_size}{Minimum number of cells in a community}

\item{phenotypes_of_interest}{Vector of phenotypes to consider}
}
\value{
A data.frame and a plot is returned
}
\description{
Identifies communities of cells based on their location. It excludes cells without a phenotype
}
