% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_no_markers.R
\docType{data}
\name{image_no_markers}
\alias{image_no_markers}
\title{SCE object of a formatted image without marker intensities (simulated by
`spaSim` package)}
\format{
An sce object. colData contains 4951 rows (cells) and 4 columns
  (metadata).
}
\usage{
image_no_markers
}
\description{
A dataset that contains a formatted sce object with cell ids, cell positions,
and cell types in metadata. This dataset does not contain assays (marker
intensities).
}
\seealso{
\code{\link{defined_image}} \code{\link{simulated_image}}
}
\keyword{datasets}
