% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_expression_boxplot.R
\name{marker_expression_boxplot}
\alias{marker_expression_boxplot}
\title{marker_expression_boxplot}
\usage{
marker_expression_boxplot(sce_object, marker)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{marker}{Marker being queried}
}
\description{
Produces boxplots of marker levels for cells phenotyped as being positive
for the marker, and those that where phenotyped as being negative.
}
