% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_intensity_boxplot.R
\name{marker_intensity_boxplot}
\alias{marker_intensity_boxplot}
\title{marker_intensity_boxplot}
\usage{
marker_intensity_boxplot(spe_object, marker)
}
\arguments{
\item{spe_object}{SpatialExperiment object in the form of the output of
\code{\link{format_image_to_spe}}.}

\item{marker}{String. Marker being queried.}
}
\value{
A plot is returned
}
\description{
Produces boxplots of marker levels for cells phenotyped as being
  positive for the marker, and those that where phenotyped as being negative.
}
\examples{
marker_intensity_boxplot(SPIAT::simulated_image, "Immune_marker1")
}
