% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_permutation.R
\name{marker_permutation}
\alias{marker_permutation}
\title{marker_permutation}
\usage{
marker_permutation(sce_object, num_iter)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of output from format_image_to_sce}

\item{num_iter}{Integer specifying the number of iterations for bootstrapping}
}
\description{
Creates random combinations of phenotypes by shuffling markers and
calculates the enrichment and depletion p values
}
