% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_prediction_plot.R
\name{marker_prediction_plot}
\alias{marker_prediction_plot}
\title{marker_prediction_plot}
\usage{
marker_prediction_plot(predicted_data, marker)
}
\arguments{
\item{predicted_data}{Output from \code{\link{predict_phenotypes}}.}

\item{marker}{String. Marker to plot}
}
\value{
A plot is returned
}
\description{
Takes in the returned dataframe from marker_threshold_plot and
generates a .pdf file containing scatter plots of actual intensity and
predicted intensity for every marker.
}
\examples{
predicted_result <- predict_phenotypes(sce_object = simulated_image, thresholds = NULL,
tumour_marker = "Tumour_marker",baseline_markers = c("Immune_marker1", "Immune_marker2", 
"Immune_marker3", "Immune_marker4"), reference_phenotypes = TRUE)
marker_prediction_plot(predicted_result, marker = "Tumour_marker")
}
