% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing_score_multiple.R
\name{mixing_score_multiple}
\alias{mixing_score_multiple}
\title{mixing_score_multiple}
\usage{
mixing_score_multiple(
  sce_object,
  reference_marker,
  target_marker,
  radius = 20,
  column = "Phenotype"
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{reference_marker}{Markers to be used as the reference cells.}

\item{target_marker}{Markers to be used as the target markers. The function will ignore instances where reference marker is the same as target marker.}

\item{radius}{(OPTIONAL) The maximum radius around a reference marker for another cell to be considered an interaction.}

\item{column}{String specifying the column where the selected cell types are from}
}
\value{
A list of dataframes
}
\description{
Produces a list of dataframes that contains all the mixing scores respective to each reference cell. Each dataframe is corresponding to one reference cell type
}
