% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morisita_index.R
\name{morisita_index}
\alias{morisita_index}
\title{morisita_index}
\usage{
morisita_index(sce_objects, CI = 0.95)
}
\arguments{
\item{sce_objects}{A vector of SingleCellExperiment object names in the form of the output of format_image_to_sce}

\item{CI}{Confidence interval for calculating morisita index}
}
\description{
Calculate the morisita index between different formatted single cell experiment images
}
