% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_of_cells_within_radius.R
\name{percentage_of_cells_within_radius}
\alias{percentage_of_cells_within_radius}
\title{percentage_of_cells_within_radius}
\usage{
percentage_of_cells_within_radius(
  sce_object,
  reference_phenotypes,
  target_phenotypes,
  radius = 100
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{reference_phenotypes}{String specifying the phenotypes of reference cells}

\item{target_phenotypes}{String specifying the phenotypes for target cells}

\item{radius}{Integer specifying the radius of search for cells around the reference cells.
Radii of ~100 are recommended. If too small, too few cells might be present}
}
\description{
Calculates the percentage of cells of a target phenotype within a radius
from the cells with a reference phenotype.
The calculation is done per reference cell, so runtime will depend on the number of reference
cells present.
}
\examples{
p_cells_within_radius <- percentage_of_cells_within_radius(SPIAT::formatted_image, reference_phenotypes = "PDL-1", target_phenotypes = "AMACR", radius=100)
}
