% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_average_expression.R
\name{plot_average_intensity}
\alias{plot_average_intensity}
\title{plot_average_intensity}
\usage{
plot_average_intensity(sce_object, reference_marker, target_marker, radii)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{reference_marker}{String specifying the reference marker}

\item{target_marker}{String specifying the marker to calculate its average intensity}

\item{radii}{Vector of integers specifying the search radius around reference cells}
}
\description{
Takes in a vector or radii and calculates the average intensity
of a target marker using average_intensity function. It plots the intensity
level as a line graph
}
