% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_categories.R
\name{plot_cell_categories}
\alias{plot_cell_categories}
\title{plot_cell_categories}
\usage{
plot_cell_categories(sce_object, phenotypes_of_interest, colour_vector)
}
\arguments{
\item{sce_object}{Singlecellexperiment object in the form of the output of format_image_to_sce}

\item{phenotypes_of_interest}{Vector of cell phenotypes to be coloured}

\item{colour_vector}{Vector specifying the colours of each cell phenotype}
}
\description{
Produces a scatter plot of the cells in the tissue. Cells are coloured
categorically by phenotype. Cells not part of the phenotypes of interest will be coloured "lightgrey"
}
\examples{
phenotypes_of_interest <- c("AMACR", "CD3,CD8", "PDL-1")
colour_vector <- c("darkgrey", "blue", "red")
plot_cell_categories(SPIAT::formatted_image, phenotypes_of_interest, colour_vector)
}
