% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_distances_violin.R
\name{plot_cell_distances_violin}
\alias{plot_cell_distances_violin}
\title{plot_cell_distances_violin}
\usage{
plot_cell_distances_violin(cell_to_cell_dist)
}
\arguments{
\item{cell_to_cell_dist}{Output from calculate_all_distances_between_phenotypes}
}
\description{
Plots distances between cells as a violin plot
}
