% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_marker_levels.R
\name{plot_cell_marker_levels}
\alias{plot_cell_marker_levels}
\title{plot_cell_marker_levels}
\usage{
plot_cell_marker_levels(sce_object, marker)
}
\arguments{
\item{sce_object}{Singlecellexperiment object in the form of the output of format_image_to_sce}

\item{marker}{Marker to plot}
}
\description{
Produces a scatter plot of the level of a marker in each cell.
Cells that were not phenotyped as being positive for the particular marker are excluded.
}
