% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_percentages.R
\name{plot_cell_percentages}
\alias{plot_cell_percentages}
\title{plot_cell_proportions}
\usage{
plot_cell_percentages(
  cell_proportions,
  tumour_marker = NULL,
  cellprop_colname = "Proportion_name"
)
}
\arguments{
\item{cell_proportions}{Output from calculate_cell_proportions}

\item{tumour_marker}{Tumour marker to exclude if needed}

\item{cellprop_colname}{Column to use for y axis names. Default is "Proportion_name"}
}
\value{
A plot is returned
}
\description{
Plots cells proportions as barplots
}
\examples{
p_cells <- calculate_cell_proportions(SPIAT::formatted_image, reference_celltypes=c("Total", "CD3"))
plot_cell_percentages(p_cells)
}
