% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marker_level_heatmap.R
\name{plot_marker_level_heatmap}
\alias{plot_marker_level_heatmap}
\title{plot_marker_level_heatmap}
\usage{
plot_marker_level_heatmap(sce_object, num_splits, marker)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{num_splits}{Integer specifying the blurring level (number of splits) for the image.
Higher numbers result in higher resolution.}

\item{marker}{Marker to plot}
}
\value{
A plot is returned
}
\description{
Blurres the image by splitting the images into small squares.
The marker levels are then averaged within each square. All cells are considered,
regardless of phenotype status
}
\examples{
plot_marker_level_heatmap(SPIAT::simulated_image, num_splits = 100, "Tumour_marker")
}
