% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mixing_score_summary.R
\name{plot_mixing_score_summary}
\alias{plot_mixing_score_summary}
\title{plot_mixing_score_summary}
\usage{
plot_mixing_score_summary(mixing_score_summary, individual.plots = FALSE)
}
\arguments{
\item{mixing_score_summary}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{individual.plots}{(OPTIONAL) Individual plots for each reference cell are given, rather than being arranged in a grid.}
}
\value{
Various plots of log2(mixing score) vs log2(number of target cells) for each reference cell.
}
\description{
Plots mixing scores against target cells from the input mixing_score_summary data.frame.
It produces multiple scatterplots for each reference cell observed, with line of best fit. 
It gives a Rho statistic with p-value based off Spearman's correlation.
It produces different sets of plots based off mixing scores and normalised mixing scores.
}
