% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_phenotypes.R
\name{predict_phenotypes}
\alias{predict_phenotypes}
\title{predict_phenotypes}
\usage{
predict_phenotypes(
  sce_object,
  thresholds = NULL,
  tumour_marker,
  baseline_markers
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{thresholds}{(Optional) Vector of numbers specifying the cutoff of a positive reading.
The order must match the marker order, and it should be NA for DAPI.}

\item{tumour_marker}{String containing the tumour_marker used for the image.}

\item{baseline_markers}{Markers not found on tumour cells to refine the threshold}
}
\description{
Produces a density plot showing predicted cutoff of a
positive reading for marker intensity. It also prints to the console the
number of true positives (TP), true negatives (TN), false positives (FP) and
false negatives (FN) under the prediction. It returns a dataframe containing
the predicted intensity status for a particular marker,
}
\examples{
predicted_image <- predict_phenotypes(SPIAT::formatted_image,
                                      thresholds = NULL,
                                      tumour_marker = "AMACR",
                                      baseline_markers = c("CD3", "CD4", "CD8"))
}
