% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_column.R
\name{print_column}
\alias{print_column}
\title{print_column}
\usage{
print_column(sce_object, column)
}
\arguments{
\item{sce_object}{SingleCellExperiment object generated by format_image_to_sce}

\item{column}{String of the specified column}
}
\value{
A character vector is returned
}
\description{
Print the unique values of a specified column
}
