% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_feature.R
\name{print_feature}
\alias{print_feature}
\title{print_feature}
\usage{
print_feature(sce_object, feature_colname)
}
\arguments{
\item{sce_object}{SingleCellExperiment object generated by
format_image_to_sce.}

\item{feature_colname}{String of the specified column.}
}
\value{
A character vector is returned
}
\description{
Print the unique values of a specified column (feature).
}
\examples{
feature_colname <- "Phenotype"
print_phenotypes(SPIAT::formatted_image, feature_colname)
}
