% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_phenotypes.R
\name{print_phenotypes}
\alias{print_phenotypes}
\title{print_phenotypes}
\usage{
print_phenotypes(sce_object)
}
\arguments{
\item{sce_object}{SingleCellExperiment object generated by format_image_to_sce}
}
\value{
A character vector is returned
}
\description{
Prints phenotypes that are present in the sample
}
\examples{
print_phenotypes(SPIAT::formatted_image)
}
