% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_BT.R
\name{R_BT}
\alias{R_BT}
\title{R_BT}
\usage{
R_BT(sce_object, cell_types_of_interest, feature_colname)
}
\arguments{
\item{sce_object}{}

\item{cell_types_of_interest}{}

\item{feature_colname}{}
}
\value{

}
\description{
R_BT
}
