% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_celltypes.R
\name{select_celltypes}
\alias{select_celltypes}
\title{select_celltypes}
\usage{
select_celltypes(
  sce_object,
  celltypes,
  feature_colname = "Phenotype",
  keep = TRUE
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object generated by
format_image_to_sce.}

\item{celltypes}{String Vector of celltypes of keep or exclude.}

\item{feature_colname}{String. The column that has the interested cell types.}

\item{keep}{Boolean. TRUE if vector of `celltypes` are the cells that are
going to be kept, FALSE if they are to be removed.}
}
\value{
A SingleCellExperiment object is returned. The original image size
  and cell count can be accessed by `attr(slim_sce, "original_cell_number")`
  and `attr(slim_sce, "range_of_coords")`.
}
\description{
Select cell types to keep or exclude in the analysis. This
  function remembers the original image size and cell count.
}
\examples{
data_subset <- select_celltypes(SPIAT::formatted_image,
                                celltypes = c("AMACR","CD3,CD8","PDL-1"),
                                feature_colname = "Phenotype", keep=TRUE)
}
