% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_phenotypes.R
\name{select_phenotypes}
\alias{select_phenotypes}
\title{select_phenotypes}
\usage{
select_phenotypes(sce_object, keep = TRUE, phenotypes = NULL)
}
\arguments{
\item{sce_object}{SingleCellExperiment object generated by format_image_to_sce}

\item{keep}{TRUE if vector of phenotypes are the cells that are going to be kept,
FALSE if they are to be removed}

\item{phenotypes}{Vector of phenotypes of keep or exclude}
}
\description{
Select phenotypes to keep or exclude in the analysis
}
