% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_image.R
\docType{data}
\name{simulated_image}
\alias{simulated_image}
\title{SCE object of a formatted image (simulated by `spaSim` package)}
\format{
An sce object. Assay contains 5 rows (markers) and 4951 columns
  (cells); colData contains 4951 rows (cells) and 3 columns (metadata).
}
\usage{
simulated_image
}
\description{
A dataset that contains a formatted sce object with cell ids, cell positions,
and phenotypes in metadata and marker intensities in assays.
}
\seealso{
\code{\link{defined_image}} \code{\link{image_no_markers}}
}
\keyword{datasets}
