% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIAT.R
\docType{package}
\name{SPIAT}
\alias{SPIAT}
\title{SPIAT: A package for spatial image analysis of tissues}
\description{
SPIAT includes six modules including reading sptial data, basic analysis and
visualization, calculation of cell colocalization, the automated
characterization of clinically relevant immune spatial patterns, measuring
spatial heterogeneity and neighborhood analysis.
}
