% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_of_cross_function.R
\name{AUC_of_cross_function}
\alias{AUC_of_cross_function}
\title{The difference in AUC of the cross function curves}
\usage{
AUC_of_cross_function(df.cross)
}
\arguments{
\item{df.cross}{Data.Frame. The output the the cross functions. Containing
the positions of the two curves.}
}
\description{
Calculate the difference of area under the curve (AUC) between
  two curves, normalised by the total area of the graph.
}
