% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_of_cross_function.R
\name{AUC_of_cross_function}
\alias{AUC_of_cross_function}
\title{AUC_of_cross_function}
\usage{
AUC_of_cross_function(df.cross)
}
\arguments{
\item{df.cross}{Dataframe containing the positions of the two curves}
}
\description{
Calculate the difference of area under curve between two curves
}
