% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_minimum_distance.R
\name{average_minimum_distance}
\alias{average_minimum_distance}
\title{average_minimum_distance}
\usage{
average_minimum_distance(sce_object)
}
\arguments{
\item{sce_object}{Singlecellexperiment object in the form of the output of format_image_to_sce}
}
\value{
A single number is returned
}
\description{
Calculates the average minimum distance of all cells in the sce_object
}
\examples{
average_minimum_distance(SPIAT::formatted_image)
}
