% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_percentage_of_cells_within_radius.R
\name{average_percentage_of_cells_within_radius}
\alias{average_percentage_of_cells_within_radius}
\title{average_percentage_of_cells_within_radius}
\usage{
average_percentage_of_cells_within_radius(
  sce_object,
  reference_phenotypes,
  target_phenotypes,
  radius = 100,
  column
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{reference_phenotypes}{String specifying the phenotypes of reference cells}

\item{target_phenotypes}{String specifying the phenotypes for target cells}

\item{radius}{Integer specifying the radius of search for cells around the reference cells.
Radii of ~100 are recommended. If too small, too few cells might be present}

\item{column}{String specifying the column with the desired cell type annotations}
}
\value{
A numeric vector ard a plot are returned
}
\description{
Calculates the average percentage of cells of a target phenotype within a radius
from the cells with a reference phenotype.
The calculation is done per reference cell, so runtime will depend on the number of reference
cells present. Output is a single value (the mean for the image).
}
