% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_dependence.R
\name{bivariate_dependence}
\alias{bivariate_dependence}
\title{bivariate_dependence}
\usage{
bivariate_dependence(
  sce_object,
  method = "Kcross",
  phenotypes,
  column,
  plot_results = TRUE
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{method}{String that is the method for dependence calculation}

\item{phenotypes}{Vector of phenotypes of interest}

\item{column}{String that is the name of the column of the types}

\item{plot_results}{TRUE if result to be plotted, FALSE if not. In either case, an object with the results is returned}
}
\description{
find the dependence between two types of cells
}
