% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cross_functions.R
\name{calculate_cross_functions}
\alias{calculate_cross_functions}
\title{calculate_cross_functions}
\usage{
calculate_cross_functions(sce_object, method = "Kcross", phenotypes, column)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{method}{String that is the method for dependence calculation. Options: "Gcross", "Kcross", "Lcross", "Jcross"}

\item{phenotypes}{Vector of phenotypes of interest}

\item{column}{String that is the name of the column of the types}
}
\description{
calculate the cross functions between two cell types to see if there is mixing between them
}
