% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_minimum_distances.R
\name{calculate_minimum_distances}
\alias{calculate_minimum_distances}
\title{calculate_minimum_distances}
\usage{
calculate_minimum_distances(
  sce_object,
  column = "Phenotype",
  cell_types_of_interest = NULL
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{column}{Column of cells to choose the phenotype from (e.g. Cell.Type, Cell.Type2, etc)}

\item{cell_types_of_interest}{Vector of marker combinations to consider
is FALSE}
}
\value{
A data.frame is returned
}
\description{
Returns the distance of the closest cell of a specific type from each reference cell
}
