% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_percentage_of_grids.R
\name{calculate_percentage_of_grids}
\alias{calculate_percentage_of_grids}
\title{calculate_percentage_of_grids}
\usage{
calculate_percentage_of_grids(raster_obj, threshold, above)
}
\arguments{
\item{raster_obj}{Raster object in the form of the output of raster function.}

\item{threshold}{Numeric. The threshold for defining the pattern.}

\item{above}{Boolean. Indicating whether the pattern is above (TRUE) or below
(FALSE) the threshold.}
}
\description{
Calculate the percentage of the grid squares that have a certain
  pattern.
}
