% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_proportions_of_cells_in_structure.R
\name{calculate_proportions_of_cells_in_structure}
\alias{calculate_proportions_of_cells_in_structure}
\title{calculate_proportions_of_cells_in_structure}
\usage{
calculate_proportions_of_cells_in_structure(sce_object)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}
}
\description{
Calculate the proprtion of immune cells in each defined tumour structure, return a dataframe
}
