% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spatial_autocorrelation.R
\name{calculate_spatial_autocorrelation}
\alias{calculate_spatial_autocorrelation}
\title{calculate_spatial_autocorrelation}
\usage{
calculate_spatial_autocorrelation(raster_obj, metric = "globalmoran")
}
\arguments{
\item{raster_obj}{Raster object in the form of the output of raster function.}

\item{metric}{String. The method for calculating spatial autocorrelation.
Choose from "globalmoran" and "GearyC". -- change code so that misspelling "GearyC" doesn't result in "globalmoran"}
}
\description{
Calculate the spatial autocorrelation metric of the grid squares
  that have a certain pattern. -- I don't understand what a 'certain pattern' is, but maybe most people would find this obvious?
}
\examples{
grid <- grid_metrics(SPIAT::defined_image, FUN = calculate_entropy, n_split = 5,
cell_types_of_interest=c("Tumour","Immune3"), feature_colname = "Cell.Type")
calculate_spatial_autocorrelation(grid, metric = "globalmoran")
}
