% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_summary_distances_of_cells_to_borders.R
\name{calculate_summary_distances_of_cells_to_borders}
\alias{calculate_summary_distances_of_cells_to_borders}
\title{calculate_summary_distances_of_cells_to_borders}
\usage{
calculate_summary_distances_of_cells_to_borders(
  sce_object,
  cell_types_of_interest,
  feature_colname = "Cell.Type"
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of
format_image_to_sce.}

\item{cell_types_of_interest}{String Vector of cell types to consider.}

\item{feature_colname}{String specifying which column the interested cell
types are from.}
}
\value{
A data.frame is returned
}
\description{
Returns the mean, median and stardard deviation of the distances
  between a specified cell type to the borders.
}
\examples{
sce_border <- identify_bordering_cells(SPIAT::defined_image, reference_cell = "Tumour",
feature_colname = "Cell.Type", n_to_exclude = 10)
sce_dist <- calculate_distance_to_tumour_margin(sce_border)
sce_structure <- define_structure(sce_dist, names_of_immune_cells = c("Immune1","Immune2","Immune3"),
feature_colname = "Cell.Type", n_margin_layers = 5)
calculate_summary_distances_of_cells_to_borders(sce_structure, 
cell_types_of_interest = c("Immune1","Immune3"),feature_colname = "Cell.Type")
}
