% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_of_clusters_and_communities.R
\name{composition_of_clusters_and_communities}
\alias{composition_of_clusters_and_communities}
\title{composition_of_clusters_and_communities}
\usage{
composition_of_clusters_and_communities(
  formatted_data_with_clusters,
  type_of_aggregate,
  column
)
}
\arguments{
\item{formatted_data_with_clusters}{- a dataframe output from generate_clusters}

\item{type_of_aggregate}{Cluster or Community}

\item{column}{Column with cell types}
}
\value{
A data.frame is returned
}
\description{
Returns a dataframe which contains the percentages of cells with a
specific marker within each cluster and the number of cells in the cluster.
}
\examples{
communities <- identify_cell_communities(SPIAT::formatted_image, radius=100)
communities_vis <- composition_of_clusters_and_communities(communities, type_of_aggregate = "Community", column="Phenotype")
}
