% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_of_neighborhoods.R
\name{composition_of_neighborhoods}
\alias{composition_of_neighborhoods}
\title{composition_of_neighborhoods}
\usage{
composition_of_neighborhoods(formatted_data_with_clusters, feature_colname)
}
\arguments{
\item{formatted_data_with_clusters}{- a dataframe output from
generate_clusters}

\item{feature_colname}{Column with cell types}
}
\value{
A data.frame is returned
}
\description{
Returns a dataframe which contains the percentages of cells with
  a specific marker within each cluster and the number of cells in the
  cluster.
}
\examples{
communities <- identify_neighborhoods(SPIAT::formatted_image, radius=100)
communities_vis <- composition_of_neighborhoods(communities, feature_colname="Phenotype")
}
