% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_celltypes.R
\name{define_celltypes}
\alias{define_celltypes}
\title{define_celltypes}
\usage{
define_celltypes(
  sce_object,
  categories = NULL,
  category_colname = "Phenotype",
  names = NULL,
  new_colname = "Cell.Type",
  print.names = F
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object generated by format_image_to_sce}

\item{categories}{Vector of categories to be defined; if NULL, the function will use 
predefined categories and names}

\item{category_colname}{(Optional) String specifying the name of the column 
having the categories to be defined, by default "Phenotype"}

\item{names}{Vector of new names assigned to the selected categories; if NULL, 
the function will use predefined categories and names}

\item{new_colname}{(Optional) String specifying the name of the column to be added, 
by default "Cell.Type"}

\item{print.names}{(Optional) Boolean if the user wants the original and new 
names printed. Default is FALSE.}
}
\description{
Define celltypes of selected categories. If the specified category vector is 
shorter than the categories in the sce_object, then the unspecified categories will be defined
as "Undefined"
}
\examples{
defined_sce <- define_celltypes(SPIAT::simulated_image, 
categories = c("Tumour_marker", "Immune_marker1,Immune_marker2", "Immune_marker1,Immune_marker3", 
"Immune_marker1,Immune_marker2,Immune_marker4", "OTHER"), category_colname = "Phenotype",
names = c("Tumour", "Immune1", "Immune2","Immune3", "Others"), new_colname = "Cell.Type")
}
