% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_structure.R
\name{define_structure}
\alias{define_structure}
\title{define_structure}
\usage{
define_structure(
  sce_object,
  names_of_immune_cells,
  feature_colname = "Cell.Type",
  n_margin_layers = 5
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object whose metadata has the
information of tumour bordering cells and cell distances to tumour border
(has columns "Region" and "Distance.To.Border").}

\item{names_of_immune_cells}{String Vector of the names of immune cells.}

\item{feature_colname}{String Specifying the column that contains the names
of the immune cells.}

\item{n_margin_layers}{Integer. The number of layers of cells that compose
the internal/external tumour margins.}
}
\value{
A new sce object is returned
}
\description{
After identifying the bordering cells of tumour regions and
  calculating the distances of each cell to the tumour bordering cells, this
  function further identifies the cells that are located in the inside and
  outside of the tumour regions, and in the internal and external tumour
  margins. It also identifies the immune cells that are infiltrated, stromal,
  internal margin or external margin immune cells.
}
\examples{
sce_border <- identify_bordering_cells(SPIAT::defined_image,
reference_cell = "Tumour", feature_colname = "Cell.Type", n_to_exclude = 10)
sce_dist <- calculate_distance_to_tumour_margin(sce_border)
sce_structure <- define_structure(sce_dist,
names_of_immune_cells = c("Immune1","Immune2","Immune3"),
feature_colname = "Cell.Type", n_margin_layers = 5)
plot_cell_categories(sce_structure, feature_colname = "Structure")
}
