% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_structure.R
\name{define_structure}
\alias{define_structure}
\title{define_structure}
\usage{
define_structure(
  sce_object,
  names_of_immune_cells,
  feature_colname = "Cell.Type",
  n_margin_layers = 5
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{names_of_immune_cells}{Vector indicating the names of potential immune cells}

\item{feature_colname}{String Specifying which column the names of immune cells are under}

\item{n_margin_layers}{Integer Specifying the number of layers of cells that compose the internal/external margins}
}
\description{
Identify the cells that compose the invasive front, the infiltration and the exclusion -- elaborate on what these mean? Or cite your paper for definitions?
}
\examples{
sce_border <- identify_bordering_cells(SPIAT::defined_image, 
reference_cell = "Tumour", feature_colname = "Cell.Type", n_to_exclude = 10)
sce_dist <- calculate_distance_to_tumour_margin(sce_border)
sce_structure <- define_structure(sce_dist, 
names_of_immune_cells = c("Immune1","Immune2","Immune3"),
feature_colname = "Cell.Type", n_margin_layers = 5)
plot_cell_categories(sce_structure, feature_colname = "Structure")
}
