% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_structure.R
\name{define_structure}
\alias{define_structure}
\title{define_structure}
\usage{
define_structure(sce_object, names_of_immune_cells, n_margin_layers = 5)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{names_of_immune_cells}{Vector indicating the names of potential immune cells}

\item{n_margin_layers}{Integer Specifying the number of layers of cells that compose the internal/external margins}
}
\description{
Identify the cells that compose the invasive front, the infiltration and the exclusion
}
