% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defined_image.R
\docType{data}
\name{defined_image}
\alias{defined_image}
\title{SPE object of a simulated image with defined cell types based on marker
combinations.}
\format{
An spe object. Assay contains 5 rows (markers) and 4951 columns
  (cells); colData contains 4951 rows (cells) and 3 columns (features).
}
\usage{
defined_image
}
\description{
A dataset that contains a formatted spe object with cell ids, phenotypes,
defined cell types in `colData()` and marker intensities in `assays()`. (The
cell locations are the same with the cells in \code{\link{simulated_image}}).
}
\seealso{
\code{\link{simulated_image}} \code{\link{image_no_markers}}
}
\keyword{datasets}
