% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cellprofiler_to_spe.R
\name{format_cellprofiler_to_spe}
\alias{format_cellprofiler_to_spe}
\title{Format an image into a SpatialExperiment object}
\usage{
format_cellprofiler_to_spe(
  path = NULL,
  markers = NULL,
  intensity_columns_interest = NULL
)
}
\arguments{
\item{path}{String of the path location cellprofiler csv file.}

\item{markers}{String Vector containing the markers used for staining.}

\item{intensity_columns_interest}{String Vector with the names of the columns
with the level of each marker. Column names must match the order of the
'markers' parameter.}
}
\value{
A SingleCellExperiment object is returned
}
\description{
Reads in spatial data in the form of cell coordinates, cell
  phenotypes (if available), and marker intensities and transforms to a
  SingleCellExperiment object. Users can input data in a general format
  (recommended), or data generated from inForm, HALO, Visum, CODEX,
  cellprofiler. The count assay stores the intensity level of every marker
  (rows) for every cell (columns). Cell phenotype, x and y coordinates and
  other cell properties are stored under colData. For the INFORM format, the
  cell properties needed to be included are Cell.Area, Nucleus.Area,
  Nucleus.Compactness, Nucleus.Axis.Ratio, and Cell.Axis.Ratio. For the HALO
  format, the cell properties to be included are Cell.Area, Nucleus.Area and
  Cytoplasm.Area. Note that if the data does not include these parameters, we
  recommend adding it to the output from inForm or HALO with NAs in columns,
  or alternatively using the ‘generic formatting’ option.
}
