% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_colData_to_sce.R
\name{format_colData_to_sce}
\alias{format_colData_to_sce}
\title{format_colData_to_sce}
\usage{
format_colData_to_sce(colData)
}
\arguments{
\item{colData}{Dataframe that will be the colData of the sce object}
}
\description{
Format a dataframe of colData into a singlecellexperiment class 
where the count assay is empty
every cell (columns), and cell phenotype, x and y coordinates are stored under colData
for the purpose of passing dataframe into a function requiring sce_object
}
