% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_image_to_sce.R
\name{format_image_to_sce}
\alias{format_image_to_sce}
\title{Format an image into a SingleCellExperiment object}
\usage{
format_image_to_sce(
  format = "INFORM",
  path = NULL,
  markers = NULL,
  locations = NULL,
  dye_columns_interest = NULL,
  intensity_columns_interest = NULL,
  path_to_visium_coordinates = NULL,
  path_to_codex_cell_phenotypes = NULL,
  intensity_matrix = NULL,
  phenotypes = NULL,
  coord_x = NULL,
  coord_y = NULL
)
}
\arguments{
\item{format}{String defining the software used for cell segmentation.
Options: "INFORM", "HALO", "Visium", "CODEX" "cellprofiler, or "general".}

\item{path}{String of the path location of either HALO csv file, INFORM text
file or CODEX csv file. If 10X Visium transcriptomics data, path to the
"filtered_feature_bc_matrix" folder generated by CellRanger. NULL is
`format` is "general".}

\item{markers}{For INFORM and HALO formats. Vector containing the markers
used for staining. Must match the order of the 'markers' parameter}

\item{locations}{Vector containing the locations of markers used for
staining. Location can be either "Nucleus", "Cytoplasm" or "Membrane". This
is used to select the Intensity column and can be used instead of
intensity_columns_interest.}

\item{dye_columns_interest}{For HALO formats. Use if locations is not
specified. Vector of names of the columns with the marker status (i.e.
those indicating 1 or 0 for whether the cell is positive or negative for
the marker). Column names must match the order of the 'markers' parameter.}

\item{intensity_columns_interest}{For HALO formats. Use if locations is not
specified. Vector with the names of the columns with the level of each
marker. Column names must match the order of the 'markers' parameter.}

\item{path_to_visium_coordinates}{For 10X Visium data. String of the path to
the "tissue_positions_list.csv" generated by CellRanger.}

\item{path_to_codex_cell_phenotypes}{For CODEX data. String of the path to
the Cluster ID/Cell type file.}

\item{intensity_matrix}{For the "general" format. A matrix of marker
intensities or gene expression where the column names are the Cell IDs, and
the rownames the marker or genes.}

\item{phenotypes}{String Vector of cell phenotypes in the same order in which
they appear in intensity_matrix. If no phenotypes available, then a vector
of NA can be used as input. Note that the combination of markers (e.g.
CD3,CD4) needs to be used instead of the cell type name (e.g. helper T
cells)}

\item{coord_x}{Numeric Vector with the X coordinates of the cells. The cells
must be in the same order as in the intensity_matrix.}

\item{coord_y}{Numeric Vector with the Y coordinates of the cells. The cells
must be in the same order as in the intensity_matrix.}
}
\value{
A SingleCellExperiment object is returned
}
\description{
Reads in spatial data in the form of cell coordinates, cell
  phenotypes (if available), and marker intensities and transforms to a
  SingleCellExperiment object. Users can input data in a general format
  (recommended), or data generated from inForm, HALO, Visum, CODEX,
  cellprofiler. The count assay stores the intensity level of every marker
  (rows) for every cell (columns). Cell phenotype, x and y coordinates and
  other cell properties are stored under colData. For the INFORM format, the
  cell properties needed to be included are Cell.Area, Nucleus.Area,
  Nucleus.Compactness, Nucleus.Axis.Ratio, and Cell.Axis.Ratio. For the HALO
  format, the cell properties to be included are Cell.Area, Nucleus.Area and
  Cytoplasm.Area. Note that if the data does not include these parameters, we
  recommend adding it to the output from inForm or HALO with NAs in columns,
  or alternatively using the ‘generic formatting’ option.
}
\examples{
# This is an example for general format
# Construct a dummy marker intensity matrix (rows are markers, columns are cells)
intensity_matrix <- matrix(c(14.557, 0.169, 1.655, 0.054, 17.588, 0.229, 1.188,
2.074, 21.262, 4.206,  5.924, 0.021), nrow = 4, ncol = 3)
# define marker names as rownames
rownames(intensity_matrix) <- c("DAPI", "CD3", "CD4", "AMACR")
# define cell IDs as colnames
colnames(intensity_matrix) <- c("Cell_1", "Cell_2", "Cell_3")
# Construct a dummy metadata (phenotypes, x/y coordinates)
# the order of the elements in these vectors correspond to the cell order
# in `intensity matrix`
phenotypes <- c("OTHER",  "AMACR", "CD3,CD4")
coord_x <- c(82, 171, 184)
coord_y <- c(30, 22, 38)

general_format_image <- format_image_to_sce(format = "general",
intensity_matrix = intensity_matrix,phenotypes = phenotypes, coord_x = coord_x,
coord_y = coord_y)
}
