% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_image_to_sce.R
\name{format_image_to_sce}
\alias{format_image_to_sce}
\title{format_image_to_sce}
\usage{
format_image_to_sce(
  format = "INFORM",
  image,
  markers,
  dye_columns_interest = NULL,
  intensity_columns_interest
)
}
\arguments{
\item{format}{String defining the software use for cell segmentation. Options: "INFORM" or "HALO"}

\item{image}{String of the path location of either HALO csv file or INFORM textfile}

\item{markers}{Vector containing the markers used for staining}

\item{dye_columns_interest}{(Only for HALO formats) Vector of names of the columns with the marker status 
(i.e. those indicating 1 or 0 for whether the cell is positive or negative for the marker). 
Column names must match the order of the 'markers' parameter.}

\item{intensity_columns_interest}{Vector with the names of the columns with the level of each marker.
Column names must match the order of the 'markers' parameter}
}
\description{
Formats an INFORM or HALO image into a singlecellexperiment class
where the count assay stores the expression level of every marker (rows) for
every cell (columns), 
and cell phenotype, x and y coordinates, other properties (Cell Size, Nucleus Size, 
Nucleus Compactness, Nucleus Axis Ratio, Cell Axis Ratio ) are stored under colData
}
