% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_metrics.R
\name{grid_metrics}
\alias{grid_metrics}
\title{grid_metrics}
\usage{
grid_metrics(sce_object, FUN, n_split, ...)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{FUN}{Variable name specifying the metric to be calculated}

\item{n_split}{Integer specifying the number of splits for the calculation of metrics}

\item{...}{arguments of FUN}
}
\value{
a list of the metrics of all grids
}
\description{
Calculate a specified metric for each little grid in the image and plot the metrics for the grids
}
