% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_bordering_cells.R
\name{identify_bordering_cells}
\alias{identify_bordering_cells}
\title{identify_bordering_cells}
\usage{
identify_bordering_cells(
  sce_object,
  tumour_marker,
  rm_noise_radius,
  radius,
  lower_bound,
  upper_bound
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{rm_noise_radius}{Number specifying the radius used for noise. Larger numbers, more leniency}

\item{radius}{Number specifying the search radius. Larger numbers, more cells to be considered}

\item{lower_bound}{Number specifying the minumum proportion of non-marker cells in the radius of the marker population}

\item{upper_bound}{Number specifying the maximum proportion of non-marker cells in the radius of the marker population}

\item{marker}{Cells positive for this marker will be used as reference}
}
\description{
Identifies the cells bordering a group of cells of a particular phenotype
}
