% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_bordering_cells.R
\name{identify_bordering_cells}
\alias{identify_bordering_cells}
\title{identify_bordering_cells}
\usage{
identify_bordering_cells(
  sce_object,
  reference_cell,
  feature_colname = "Cell.Type",
  ahull_alpha = NULL,
  n_of_polygons = 1,
  draw = F,
  n_to_exclude = 10
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{reference_cell}{Cells positive for this marker will be used as reference}

\item{feature_colname}{Column to select for phenotypes. Can be Phenotypes, Cell.Type, etc}

\item{ahull_alpha}{Number specifying the ahull parameter. Larger number, more points included in the ahull.}

\item{n_of_polygons}{Number specifying the number of tumour regions defined by user}

\item{draw}{Boolean if user chooses to draw the tumour area or not. Default is False.}

\item{n_to_exclude}{Number Clusters under this number will be deleted}
}
\description{
Identify the cells bordering a group of cells of a particular phenotype -- needs longer explanation of how this works
}
\examples{
sce_border <- identify_bordering_cells(SPIAT::defined_image, reference_cell = "Tumour",
feature_colname = "Cell.Type", n_to_exclude = 10)
}
