% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_bordering_cells.R
\name{identify_bordering_cells}
\alias{identify_bordering_cells}
\title{identify_bordering_cells}
\usage{
identify_bordering_cells(
  sce_object,
  reference_cell,
  draw = F,
  n_of_polygons = 1,
  ahull_alpha = NULL,
  column = "Cell.Type"
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{reference_cell}{Cells positive for this marker will be used as reference}

\item{draw}{Boolean if user chooses to draw the tumour area or not. Default is False.}

\item{n_of_polygons}{Number specifying the number of tumour regions defined by user}

\item{ahull_alpha}{Number specifying the ahull parameter. Larger number, more points included in the ahull.}

\item{column}{Column to select for phenotypes. Can be Phenotypes, Cell.Type, etc}
}
\description{
Identify the cells bordering a group of cells of a particular phenotype
}
