% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_cell_clusters.R
\name{identify_cell_clusters}
\alias{identify_cell_clusters}
\title{identify_cell_clusters}
\usage{
identify_cell_clusters(
  sce_object,
  cell_types_of_interest,
  radius,
  column,
  no_pheno = NULL
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{cell_types_of_interest}{Vector of phenotypes to consider}

\item{radius}{Integer specifying the radius of search.}

\item{column}{Column from which the cell types are selected}

\item{no_pheno}{Cell type corresponding to cells without a known phenotype (e.g. "None", "Other")}
}
\value{
A data.frame and a plot is returned
}
\description{
Uses Euclidean distances to identify clusters of cells within a specified radius.
}
