% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_neighborhoods.R
\name{identify_neighborhoods}
\alias{identify_neighborhoods}
\title{identify_neighborhoods}
\usage{
identify_neighborhoods(
  sce_object,
  method = "hierarchical",
  cell_types_of_interest,
  radius,
  min_cluster_size = 10,
  k = 100,
  feature_colname,
  no_pheno = NULL
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of
format_image_to_sce.}

\item{method}{String. The clustering method. Choose from "hierarchical",
"dbscan" and "Rphenograph".}

\item{cell_types_of_interest}{String Vector of phenotypes to consider.}

\item{radius}{Numeric specifying the radius of search.}

\item{min_cluster_size}{Numeric. The minimum number of cells within each
cluster.}

\item{k}{Numeric. The parameter for Rphenograph method.}

\item{feature_colname}{String. Column from which the cell types are selected.}

\item{no_pheno}{Cell type corresponding to cells without a known phenotype
(e.g. "None", "Other")}
}
\value{
A data.frame and a plot is returned
}
\description{
Uses Euclidean distances to identify clusters of cells within a
  specified radius.
}
