% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_no_markers.R
\docType{data}
\name{image_no_markers}
\alias{image_no_markers}
\title{SPE object of a formatted image without marker intensities (simulated by
`spaSim` package)}
\format{
An spe object. colData contains 4951 rows (cells) and 3 columns
  (features).
}
\usage{
image_no_markers
}
\description{
A dataset that contains a formatted sce object with cell ids and cell types
in `colData()` and cell coordinates in `spatialCoords()`. This dataset does
not contain assays (marker intensities).
}
\seealso{
\code{\link{defined_image}} \code{\link{simulated_image}}
}
\keyword{datasets}
