% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_splitter.R
\name{image_splitter}
\alias{image_splitter}
\title{image_splitter}
\usage{
image_splitter(
  sce_object,
  number_of_splits,
  plot = FALSE,
  cut_labels = TRUE,
  colour_vector = NULL,
  x_position_min = NULL,
  x_position_max = NULL,
  y_position_min = NULL,
  y_position_max = NULL,
  column = "Cell.Type"
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{number_of_splits}{Number specifying the number of segments (e.g. 2 = 2x2, 3 = 3x3)}

\item{plot}{Specifies whether the splitted images should be printed in a pdf}

\item{cut_labels}{- specifies whether to plot where the image had been segmented}

\item{colour_vector}{List of colors. If specified, the colours will be used for plotting.
If NULL, colors will be generated automatically}

\item{x_position_min}{Integer used to specify the minimum x boundary to be splitted}

\item{x_position_max}{Integer used to specify the maximum x boundary to be splitted}

\item{y_position_min}{Integer used to specify the minimum y boundary to be splitted}

\item{y_position_max}{Integer used to specify the maximum y boundary to be splitted}

\item{column}{String specifying which column the colouring should be based on}
}
\value{
A list of data.frames is returned
}
\description{
Takes in a singlecellexperiment object from format_image_to_sce,
splits the image into specified sections and plots the different combinations
of markers within the image segments.
}
\examples{
split_image <- image_splitter(SPIAT::formatted_image, number_of_splits=3, plot = FALSE)
}
