% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_prediction_plot.R
\name{marker_prediction_plot}
\alias{marker_prediction_plot}
\title{marker_prediction_plot}
\usage{
marker_prediction_plot(predicted_data, marker)
}
\arguments{
\item{predicted_data}{Output from predict_phenotypes}

\item{marker}{Marker to plot}
}
\value{
A plot is returned
}
\description{
Takes in the returned dataframe from marker_threshold_plot and
generates a .pdf file containing scatter plots of actual intensity and
predicted intensity for every marker.
}
