% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_association_to_cell_properties.R
\name{measure_association_to_cell_properties}
\alias{measure_association_to_cell_properties}
\title{measure_association_to_cell_properties}
\usage{
measure_association_to_cell_properties(
  sce_object,
  property = "Cell.Size",
  phenotypes,
  merge = NULL,
  merge_name = NULL,
  method = "density",
  Nucleus.Ratio = FALSE,
  log.scale = FALSE
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{property}{String that is the name of the column of interest}

\item{phenotypes}{Vector of phenotypes of interest}

\item{merge}{Vector of phenotypes to be merged}

\item{merge_name}{String that is the name of the merged phenotype}

\item{method}{the analysis to do on the selected phenotypes and property}

\item{Nucleus.Ratio}{when the ratio of the nucleus size is of interest}

\item{log.scale}{if log the data}
}
\description{
plot the density or boxplot of properties of certain cell phenotypes
compare density plots or boxplots of different cell phenotypes
t test/wilcoxon rank sum test of a property of two cell phenotypes
}
