% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_cells_within_radius.R
\name{number_of_cells_within_radius}
\alias{number_of_cells_within_radius}
\title{number_of_cells_within_radius}
\usage{
number_of_cells_within_radius(
  sce_object,
  reference_marker,
  target_marker,
  radius = 20,
  column
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object in the form of the output of format_image_to_sce}

\item{reference_marker}{Marker to be used for reference cells}

\item{target_marker}{Marker to be used for target cells}

\item{radius}{Radius around the reference cells}

\item{column}{String specifying the column with the desired cell type annotations}
}
\value{
List of dataframes with the number of target cells of each of the reference cells
}
\description{
Calculates the number of cells positive for a target marker within a pre-defined radius around cells positive for a reference marker
}
