% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_basic.R
\name{plot_cell_basic}
\alias{plot_cell_basic}
\title{plot_cell_basic}
\usage{
plot_cell_basic(
  sce_object,
  types_of_interest,
  colour_vector,
  column,
  cex = 0.4
)
}
\arguments{
\item{sce_object}{Singlecellexperiment object in the form of the output of format_image_to_sce}

\item{types_of_interest}{Vector of cell cell types to be coloured}

\item{colour_vector}{Vector specifying the colours of each cell phenotype}

\item{column}{String specifying the column to be coloured}

\item{cex}{Number specifying the size of the points on the plot. Default is 0.4.}
}
\value{
A plot is returned
}
\description{
Produces a scatter plot of the cells in the tissue. Cells are coloured
categorically by specified column. Cells not part of the celltypes of interest will be coloured "lightgrey"
}
