% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_distances_violin.R
\name{plot_cell_distances_violin}
\alias{plot_cell_distances_violin}
\title{plot_cell_distances_violin}
\usage{
plot_cell_distances_violin(cell_to_cell_dist)
}
\arguments{
\item{cell_to_cell_dist}{Data.frame. Output from
\code{\link{calculate_all_distances_between_phenotypes}}.}
}
\value{
A plot is returned
}
\description{
Plots distances between cells as a violin plot
}
\examples{
distances <- calculate_distances_between_cell_types(SPIAT::defined_image,
cell_types_of_interest = c("Immune1", "Immune2"), feature_colname="Cell.Type")
plot_cell_distances_violin(distances)
}
