% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_marker_levels.R
\name{plot_cell_marker_levels}
\alias{plot_cell_marker_levels}
\title{plot_cell_marker_levels}
\usage{
plot_cell_marker_levels(
  sce_object,
  print = TRUE,
  filename = NULL,
  return_data = TRUE
)
}
\arguments{
\item{sce_object}{Singlecellexperiment object in the form of the output of format_image_to_sce}

\item{print}{TRUE for plots to be printed, FALSE otherwise}

\item{filename}{Path and name of output pdf file if to be printing to a file, if required}

\item{return_data}{TRUE if the function should return the formatted data for plotting}
}
\description{
Produces a scatter plot of the level of every marker in each cell.
Cells that were not phenotyped as being positive for the particular marker are excluded.
}
