% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_marker_levels.R
\name{plot_cell_marker_levels}
\alias{plot_cell_marker_levels}
\title{plot_cell_marker_levels}
\usage{
plot_cell_marker_levels(spe_object, marker)
}
\arguments{
\item{spe_object}{SpatialExperiment object in the form of the output of
\code{\link{format_image_to_spe}}.}

\item{marker}{String. Marker to plot.}
}
\value{
A plot is returned
}
\description{
Produces a scatter plot of the level of a marker in each cell.
  The level of the marker in all cells is shown, at x-y positions, no matter
  if cells are phenotyped as being positive or negative for the particular
  marker.
}
\examples{
plot_cell_marker_levels(SPIAT::simulated_image, "Immune_marker1")
}
