% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_percentages.R
\name{plot_cell_percentages}
\alias{plot_cell_percentages}
\title{plot_cell_proportions}
\usage{
plot_cell_percentages(
  cell_proportions,
  cells_to_exclude = NULL,
  cellprop_colname = "Proportion_name"
)
}
\arguments{
\item{cell_proportions}{Data Frame. Output from `calculate_cell_proportions`.}

\item{cellprop_colname}{Column to use for y axis names. Default is "Proportion_name"}

\item{tumour_marker}{String. Tumour marker to exclude if needed.}
}
\value{
A plot is returned
}
\description{
Plots cells proportions as barplots.
}
\examples{
p_cells <- calculate_cell_proportions(SPIAT::formatted_image, reference_celltypes=c("Total", "CD3"))
plot_cell_percentages(p_cells)
}
