% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cell_percentages.R
\name{plot_cell_percentages}
\alias{plot_cell_percentages}
\title{plot_cell_proportions}
\usage{
plot_cell_percentages(cell_proportions, tumour_marker = NULL)
}
\arguments{
\item{cell_proportions}{Output from calculate_cell_proportions}

\item{tumour_marker}{Tumour marker to exclude if needed}
}
\description{
Plots cells proportions as barplots
}
