% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_composition_heatmap.R
\name{plot_composition_heatmap}
\alias{plot_composition_heatmap}
\title{plot_composition_heatmap}
\usage{
plot_composition_heatmap(
  composition,
  pheno_to_exclude = NULL,
  log_values = FALSE,
  column_to_consider
)
}
\arguments{
\item{composition}{- a dataframe output from cluster_percent_composition}

\item{pheno_to_exclude}{Vector of phenotype to exclude}

\item{log_values}{TRUE if the percentages should be logged (base 10)}

\item{column_to_consider}{Column name to consider as community/clusters}
}
\value{
A plot is returned
}
\description{
Produces a heatmap showing the marker percentages within each cluster
and the cluster sizes
}
\examples{
communities <- identify_cell_communities(SPIAT::formatted_image, radius=100)
communities_vis <- composition_of_clusters_and_communities(communities, "Community")
plot_composition_heatmap(communities_vis, column_to_consider="Community")
}
