% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_heatmap.R
\name{plot_distance_heatmap}
\alias{plot_distance_heatmap}
\title{plot_distance_heatmap}
\usage{
plot_distance_heatmap(phenotype_distances_result, metric = "mean")
}
\arguments{
\item{phenotype_distances_result}{Dataframe output from calculate_distances_between_phenotypes}

\item{metric}{Metric to be plotted. One of "mean", "std.dev" or "median".}
}
\value{
A plot is returned
}
\description{
Takes the output of cell_distances and plot the distances as a heatmap
}
\examples{
summary_distances <- calculate_summary_distances_between_cell_types(SPIAT::formatted_image)
plot_distance_heatmap(summary_distances)
}
