% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_celltypes.R
\name{select_celltypes}
\alias{select_celltypes}
\title{select_celltypes}
\usage{
select_celltypes(
  sce_object,
  celltypes,
  feature_colname = "Phenotype",
  keep = TRUE
)
}
\arguments{
\item{sce_object}{SingleCellExperiment object generated by format_image_to_sce}

\item{celltypes}{Vector of celltypes of keep or exclude}

\item{feature_colname}{String the name of the feature column that the celltypes are under}

\item{keep}{TRUE if vector of phenotypes are the cells that are going to be kept,
FALSE if they are to be removed}
}
\value{
A SingleCellExperiment object is returned
}
\description{
Select celltypes to keep or exclude in the analysis
}
\examples{
data_subset <- select_celltypes(SPIAT::formatted_image,
                                celltypes = c("AMACR","CD3,CD8","PDL-1"), 
                                feature_colname = "Phenotype", keep=TRUE)
}
