% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPIAT-package.R
\docType{package}
\name{SPIAT-package}
\alias{SPIAT}
\alias{SPIAT-package}
\title{SPIAT: Spatial Image Analysis of Tissues}
\description{
SPIAT (**Sp**atial **I**mage **A**nalysis of **T**issues) is an R package with a suite of data processing, quality control, visualization and data analysis tools. SPIAT is compatible with data generated from single-cell spatial proteomics platforms (e.g. OPAL, CODEX, MIBI, cellprofiler). SPIAT reads spatial data in the form of X and Y coordinates of cells, marker intensities and cell phenotypes. SPIAT includes six analysis modules that allow visualization, calculation of cell colocalization, categorization of the immune microenvironment relative to tumor areas, analysis of cellular neighborhoods, and the quantification of spatial heterogeneity, providing a comprehensive toolkit for spatial data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://trigosteam.github.io/SPIAT/}
  \item Report bugs at \url{https://github.com/trigosteam/SPIAT/issues}
}

}
\author{
\strong{Maintainer}: Yuzhou  Feng \email{yuzhou.feng@petermac.org} (\href{https://orcid.org/0000-0002-2955-4053}{ORCID})

Authors:
\itemize{
  \item Anna Trigos \email{anna.trigos@petermac.org} (\href{https://orcid.org/0000-0002-5915-2952}{ORCID})
  \item Tianpei  Yang \email{yangtp3344@gmail.com}
  \item Mabel  Li
  \item John  Zhu
  \item Volkan  Ozcoban \email{volkanozcoban1@gmail.com}
  \item Maria Doyle \email{maria.doyle@petermac.org}
}

}
\keyword{internal}
